Implement correct running from battles
It is no longer guaranteed that running from a battle will work

It is sometimes impossible to run from a battle - such information is displayed when selecting an option to run. Attempting to run when it is impossible has no consequence - character's move is NOT wasted

If it is possible to run, the chance of success is dependent on a ratio of current health of player's party to their max health. The chance is displayed when selecting an option to run.

Failing to run uses up a move for the current character

Fixed displayed info for Welenc's passive ability

Fixed displayed info for Yes and No options when attempting to guard or run